[Feat] #8 : check user before deleting or updating post
Check whehther the user trying to delet or update the post
is the author of the post from the client side.

+++ client/src/pages/PostDetail.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 
 import * as api from '../lib/api';
 import usePromise from '../hooks/usePromise';
@@ -8,10 +9,15 @@ import Post from '../components/Post';
 
 const PostDetail = ({history}) => {
     const { id } = useParams();
+    const { username } = useSelector(state => state.user);
     const [loading, post, error] = usePromise(() => api.getPost(id), []);
     
     const handleCancel = async () => {
         try {
+            if (username !== post.username){
+                alert("You are not allowed to delete this post.");
+                return;
+            }
             const response = await api.deletePost(id);
 
             if (response.status === 202){
@@ -23,6 +29,10 @@ const PostDetail = ({history}) => {
     }
     
     const handleEdit = async () => {
+        if (username !== post.username){
+            alert("You are not allowed to edit this post.");
+            return;
+        }
         history.push(`/post/edit/${id}`);
     }
